+++ frontend/src/features/main/ui/main/index.ts
@@ -1,1 +1,2 @@
 export * from './MainComp'
+export * from './MainResponsive'
