+++ frontend/src/shared/components/index.ts
@@ -1,2 +1,3 @@
 export * from './typography'
 export * from './sidebar'
+export * from './Header'
